fix 오류 수정

+++ src/main/java/com/example/CafeAutoSystem/common/repository/CurrentStockLogRepository.java
@@ -6,8 +6,8 @@ import java.util.List;
 
 public interface CurrentStockLogRepository extends JpaRepository<CurrentStockLogEntity, Long> {
 
-    List<CurrentStockLogEntity> findByIngredient_IngredientId(Long ingredientId);
+    List<CurrentStockLogEntity> findByIngredient_IngredientId(Integer ingredientId);
 
     /** 재료별 로그 최신순 조회 */
-    List<CurrentStockLogEntity> findByIngredient_IngredientIdOrderByCreatedAtDesc(Long ingredientId);
+    List<CurrentStockLogEntity> findByIngredient_IngredientIdOrderByCreatedAtDesc(Integer ingredientId);
 }
